fix: refactor the employee endpoint to have a checker for organization first

+++ employees/employees.controllers.js
@@ -1,11 +1,21 @@
 import asyncHandler from "express-async-handler";
 import Employee from "./employees.models.js";
+import Organization from "../organizations/organizations.models.js";
+import Team from "../teams/teams.models.js";
 
 // Get all employees by organization
 const getAllEmployeesByOrganization = asyncHandler(async (req, res) => {
   const { organizationId } = req.params;
   try {
     const employees = await Employee.find({ organization: organizationId });
+
+    const organization = await Organization.findById(organizationId);
+    if (!organization) {
+      return res
+        .status(404)
+        .json({ error: `Organization ${organizationId} not found` });
+    }
+
     res.status(200).json(employees);
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -16,6 +26,13 @@ const getAllEmployeesByOrganization = asyncHandler(async (req, res) => {
 const getEmployeeByIdByOrganization = asyncHandler(async (req, res) => {
   const { organizationId, employeeId } = req.params;
   try {
+    const organization = await Organization.findById(organizationId);
+    if (!organization) {
+      return res
+        .status(404)
+        .json({ error: `Organization ${organizationId} not found` });
+    }
+
     const employee = await Employee.findOne({
       _id: employeeId,
       organization: organizationId,
@@ -25,6 +42,11 @@ const getEmployeeByIdByOrganization = asyncHandler(async (req, res) => {
         message: `Employee ${employeeId} not found in the organization`,
       });
     }
+
+    if (!organizationId) {
+      return res.status(404).json({ error: "Organization not found" });
+    }
+
     res.status(200).json(employee);
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -35,6 +57,13 @@ const getEmployeeByIdByOrganization = asyncHandler(async (req, res) => {
 const updateEmployeeByOrganization = asyncHandler(async (req, res) => {
   const { organizationId, employeeId } = req.params;
   try {
+    const organization = await Organization.findById(organizationId);
+    if (!organization) {
+      return res
+        .status(404)
+        .json({ error: `Organization ${organizationId} not found` });
+    }
+
     const employee = await Employee.findOneAndUpdate(
       { _id: employeeId, organization: organizationId },
       req.body,
@@ -55,6 +84,13 @@ const updateEmployeeByOrganization = asyncHandler(async (req, res) => {
 const deleteEmployeeByOrganization = asyncHandler(async (req, res) => {
   const { organizationId, employeeId } = req.params;
   try {
+    const organization = await Organization.findById(organizationId);
+    if (!organization) {
+      return res
+        .status(404)
+        .json({ error: `Organization ${organizationId} not found` });
+    }
+
     const employee = await Employee.findOneAndDelete({
       _id: employeeId,
       organization: organizationId,
@@ -78,15 +114,3 @@ export {
   updateEmployeeByOrganization,
   deleteEmployeeByOrganization,
 };
-
-// Create a new employee
-//TODO
-// delete this
-// const createEmployee = asyncHandler(async (req, res) => {
-//   try {
-//     const employee = await Employee.create(req.body);
-//     res.status(201).json(employee);
-//   } catch (error) {
-//     res.status(400).json({ error: error.message });
-//   }
-// });
